csfixed

+++ src/cart/AbstractCartPosition.php
@@ -1,12 +1,11 @@
 <?php
-
-/*
+/**
  * Finance module for HiPanel
  *
  * @link      https://github.com/hiqdev/hipanel-module-finance
  * @package   hipanel-module-finance
  * @license   BSD-3-Clause
- * @copyright Copyright (c) 2015-2016, HiQDev (http://hiqdev.com/)
+ * @copyright Copyright (c) 2015-2017, HiQDev (http://hiqdev.com/)
  */
 
 namespace hipanel\modules\finance\cart;
@@ -79,7 +78,7 @@ abstract class AbstractCartPosition extends ActiveRecord implements CartPosition
      */
     public function setPrice($price)
     {
-        $this->_price = (double) $price;
+        $this->_price = (float) $price;
     }
 
     /**
@@ -162,6 +161,6 @@ abstract class AbstractCartPosition extends ActiveRecord implements CartPosition
      */
     public function setValue($value)
     {
-        $this->_value = (double) $value;
+        $this->_value = (float) $value;
     }
 }
